Map responses to Entities
This commit introduces a special Response object that contains the
resource representation (which is an object if a mapper is defined
in the Description's model section) and the meta section.

With this mechanism it is trivial to add a mapper and have the
returned associative array transformed into a real entity.