Add event submission and change fetch to use URL
In the description for events I have added a new operation to submit new events,
this submission returns an array with one member 'url'. The url member points to
the newly created event.

In addition I have changed the `fetch` operation to use a URL as parameter; this
will allow the user to fetch a given event in HATEOAS style.